more documentation, removed CommandWrapper's Exited attribute, handle error cases, replaced literal response codes with http enums

+++ execUtil.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"errors"
 	"io"
 	"os"
 	"os/exec"
@@ -13,7 +14,7 @@ import (
 
 // TODO prune finished tasks when some max map size is reached
 
-var taskIdCommandMap SyncMap
+var taskIDCommandMap SyncMap
 var hostname string
 
 func init() {
@@ -23,48 +24,46 @@ func init() {
 		panic(err)
 	}
 
-	taskIdCommandMap = NewMap()
+	taskIDCommandMap = NewMap()
 }
 
 // GetProcessStatus is called to retrieve the details of a processes by task_id.
-func GetProcessStatus(task_id string) StatusResponse {
+func GetProcessStatus(taskID string) (*StatusResponse, error) {
 
-	var s StatusResponse
-	s.Task_id = task_id
+	var statusResponse StatusResponse
+	statusResponse.TaskID = taskID
 
 	var command *CommandWrapper
 	var ok bool
 
 	// validate task_id
-	if command, ok = taskIdCommandMap.Get(task_id); !ok {
-		s.Errors = []string{"invalid task_id"}
-		return s
+	if command, ok = taskIDCommandMap.Get(taskID); !ok {
+		return nil, errors.New("invalid task_id")
 	}
 
 	// TODO refactor this into un/finished process
-	s.Task_id = task_id
-	s.StartTime = new(time.Time)
-	*s.StartTime = command.StartTime
-	s.Finished = new(bool)
-	*s.Finished = command.GetFinished()
-	s.Errors = command.GetErrors()
-
-	// cmd.Wait() has finished, append
-	if command.GetFinished() {
-		s.EndTime = new(time.Time)
-		*s.EndTime = command.EndTime
-		s.ExitCode = new(int)
-		*s.ExitCode = command.GetExitCode()
+	statusResponse.TaskID = taskID
+	statusResponse.StartTime = new(time.Time)
+	*statusResponse.StartTime = command.StartTime
+	statusResponse.ExecError = command.GetExecError()
+	statusResponse.Finished = new(bool)
+	if command.GetEndTime() != nil {
+		*statusResponse.Finished = true
+		statusResponse.EndTime = new(time.Time)
+		statusResponse.EndTime = command.GetEndTime()
+		statusResponse.ExitCode = new(int)
+		*statusResponse.ExitCode = command.GetExitCode()
 	}
 
-	s.Output = command.StdoutBuff.Lines()
+	statusResponse.Output = command.StdoutBuff.Lines()
 
-	return s
+	return &statusResponse, nil
 }
 
 // RunCommand starts a process from command argument.
-func RunCommand(command string) StartResponse {
-	var s StartResponse
+func RunCommand(command string) (*StartResponse, error) {
+
+	var startResponse StartResponse
 
 	cmd := exec.Command(command)
 	cmd.SysProcAttr = &syscall.SysProcAttr{Setpgid: true}
@@ -75,81 +74,73 @@ func RunCommand(command string) StartResponse {
 	err := cmd.Start()
 
 	if err != nil {
-		s.Error = err.Error()
-		return s
+		return nil, err
 	}
 
 	pgid, err := syscall.Getpgid(cmd.Process.Pid)
 
 	if err != nil {
-		s.Error = err.Error()
-		return s
+		return nil, err
 	}
 
-	task_id := hostname + "-" + strconv.Itoa(pgid) // TODO handle if Process or Pid nil
-	s.Task_id = task_id
+	taskID := hostname + "-" + strconv.Itoa(pgid) // TODO handle if Process or Pid nil
+	startResponse.TaskID = taskID
 
-	taskIdCommandMap.Put(task_id, NewCommandWrapper(cmd, outBuf))
+	taskIDCommandMap.Put(taskID, NewCommandWrapper(cmd, outBuf))
 
-	// async subroutine
 	go func() {
 		// TODO add append error to CommandWrapper, impl accessors and setters
-		err = cmd.Wait()
-		cw, _ := taskIdCommandMap.Get(task_id)
-		if err != nil {
-			cw.AppendError(err.Error())
+		waitErr := cmd.Wait()
+		cw, _ := taskIDCommandMap.Get(taskID)
+		if waitErr != nil {
+			cw.SetExecError(waitErr.Error())
 		}
 		cw.SetExitCode(cmd.ProcessState.ExitCode())
-		cw.SetFinished(true)
+		cw.SetEndTime(time.Now())
 	}()
 
-	return s
+	return &startResponse, nil
+}
+
+func intPtr(value int) *int {
+	return &value
 }
 
 // StopProcess is called to end a previously started process.
-func StopProcess(task_id string) StopResponse {
+func StopProcess(taskID string) (*StopResponse, error) {
 
-	var s StopResponse
-	s.Task_id = task_id
+	var stopResponse StopResponse
+
+	stopResponse.TaskID = taskID
 
 	var command *CommandWrapper
 	var ok bool
 
 	// if task_id invalid, return error.
-	if command, ok = taskIdCommandMap.Get(task_id); !ok {
-		s.Error = []string{"invalid task_id"}
-		return s
+	if command, ok = taskIDCommandMap.Get(taskID); !ok {
+		return &stopResponse, errors.New("invalid task_id")
 	}
 
-	s.ExitCode = new(int)
-
-	if command.GetFinished() {
-		s.Error = []string{"process already finished."}
-		*s.ExitCode = command.GetExitCode()
-		return s
+	if command.GetEndTime() != nil {
+		stopResponse.ExitCode = intPtr(command.GetExitCode())
+		return &stopResponse, errors.New("process already finished")
 	}
 
 	// send stop signal first
 	// TODO verify flat processes work with gpid as well
 
-	pgid, err := syscall.Getpgid(command.Command.Process.Pid)
-	err = syscall.Kill(-pgid, syscall.SIGINT)
+	pgid, _ := syscall.Getpgid(command.Command.Process.Pid)
+	err := syscall.Kill(-pgid, syscall.SIGQUIT)
 
-	if err != nil {
-
-		errs := []string{err.Error()}
+	if err != nil && err != syscall.EPERM {
 
 		err = syscall.Kill(-pgid, syscall.SIGKILL)
 		if err != nil {
-			errs = append(errs, err.Error())
-			s.Error = errs
-			return s
+			return &stopResponse, err
 		}
-
-		s.Error = errs
 	}
 
-	*s.ExitCode = command.GetExitCode()
+	stopResponse.ExitCode = intPtr(command.GetExitCode())
 
-	return s
+	return &stopResponse, err
 }
